+++ 06-portfolio/solutions/po_a010_t10_orig.bst.sol
@@ -1,0 +1,13 @@
+# Arvak PCE dense (statevector simulation)
+# Objective value = -0.030604
+# 10 vars -> 4 qubits
+0
+0
+0
+1
+1
+1
+1
+1
+1
+1