+++ misc/ci/check_submission.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python3
 # This file is part of QOBLIB - Quantum Optimization Benchmarking Library
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.