+++ misc/tools/get_metrics.py
@@ -1,3 +1,8 @@
+#!/usr/bin/env -S uv run --script
+# /// script
+# requires-python = ">=3.12"
+# dependencies = ["pyscipopt"]
+# ///
 # This file is part of QOBLIB - Quantum Optimization Benchmarking Library
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.